chore: Set up CI workflow to test placeholder action
This commit introduces the initial CI workflow to "dogfood" our own
action.

The workflow runs on every pull request, checking out the code and
invoking the local action with test inputs. This serves as a basic
harness to validate the action.yml definition and will be expanded upon
as the tool's functionality is developed.

Resolves #14.

Signed-off-by: AlienHeadwars <bantam2781@gmail.com>