fix(env): Add rsync to Nix development environment
This commit resolves a test failure where the `rsync` command could not
be found. The `TestRunEndToEnd` test depends on the `rsync` binary to
execute the core functionality of the application.

The `rsync` package has been added to the `dev.nix` file to ensure it is
available in the development environment, allowing the tests to run
successfully.

Signed-off-by: alienHeadwars <bantam2781@gmail.com>