feat: Add stub implementation for remapper package
This commit introduces the new `internal/remapper` package and lays the
groundwork for the file extension remapping feature.

It includes the initial stub functions for `ParseExtensionMap` and
`RemapExtensions`, along with a `FileSystem` interface that will be
used for dependency injection. This allows for the creation of unit
tests that are decoupled from the actual file system.

Signed-off-by: alienHeadwars <bantam2781@gmail.com>