docs: Update README with installation and manifest details
This commit updates the `README.md` to provide a complete guide for
new users, resolving issue \#56.

The "Installation" section has been updated with the `go install`
command and a list of prerequisites. A new "Manifest File Format"
section has been added to formally document the syntax of the allow-list
file, including an example.

These changes ensure that any developer discovering the project has all
the information they need to install, configure, and use the tool
successfully.

Signed-off-by: alienHeadwars [bantam2781@gmail.com](mailto:bantam2781@gmail.com)