fix(action): Add local binary path for testing
Refactors the action to make it testable in a pull request context. A
new `local-binary-path` input has been added. If this input is
provided, the action will skip the release download step and use the
local binary instead.

This change resolves the "no assets match" error in the test workflow
by allowing the test to build and inject its own binary, decoupling the
action's test from the release process.